BUGFIX: add api user needs to return formatted output
When running ```stack add api user```, the code would print
the credentials of  the user that were just created.
This mean when this command was called through the ReST API,
it would fail, since it would print the output on standard out,
and not through stacki command line formatted output.

This commit fixes the above problem.